Add files via upload
This project delves into an extensive dataset of flight information spanning from 1960 to 2014. The primary goal is to conduct a comprehensive analysis of aviation trends, patterns, and anomalies over this extensive period. Utilizing the capabilities of Neo4J and MongoDB, two powerful and distinct database systems, the project offers a unique approach to data handling and interpretation.